Better update handling, cache optional now.
Implemented a better handling of updates, now the timestamp of the most
current changed page/media file is used as timestamp for deciding where
recent changes shall begin. Additionaly, the cache is optional now.
Use this when you have a good connectivity to your wiki and fast
updates are more important than good performance and low load on the
server side.

This version is only compatible with DokuWiki release 2009-02-14 and newer.