Use OptionParser, updater enabled for media.
DokuFS now uses OptionParser for parsing commandline arguments.
The commandline options have been changed a tiny bit and there are
new shortcuts and a (hopefully) better usage message. Generally
this means that there are better error messages now and the options
are easier to manage.

Somehow the updater hasn't yet been enabled for media files, now it is.