Fixed the updater-thread that pulls recent changes.
There were some bugs that prevented updates from actually happening.
Some were quite easy, some like the fact that the lastModified time
isn't UTC required a creative workaround. Now it is possible that
there are too many changes fetched (at maximum 24 hours), but it
should work with different time zones on client/server and with clocks
that are completely out of sync.

Furthermore there is a tiny bug in the xmlrpc interface of dokuwiki that
prevents change dates from being calculated for media files at all in the
recent changes call. There exists a workaround for this bug now, too.