Implemented a configuration file with profiles.
You can now create a configuration file that contains profiles that you
can select on the commandline. Note: The options have changed in this
version, -p stands no longer for --password but for --profile (as you
normally don't want to provide any passwords on the commandline) and the
directory is no longer the last argument but another switch. For more
information on profiles look into the help message.